[Refactor] card 모듈 추가
