--- src/main/kotlin/com/whatever/caro/card/internal/CardRepository.kt
@@ -1,5 +1,0 @@
-package com.whatever.caro.card.internal
-
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface CardRepository : JpaRepository<Card, Long>
